+++ app/controllers/pwb/props_controller.rb
@@ -9,18 +9,22 @@ module Pwb
       # gon.property_details =@property_details
       @operation_type = "for_rent"
       @operation_type_key = @operation_type.camelize(:lower)
+      @map_markers = []
       if @property_details && @property_details.visible && @property_details.for_rent
+        set_map_marker
         # below lets me know what prices to display
         @show_vacational_rental = @property_details.for_rent_short_term
 
         js property_details: @property_details
         js 'Pwb/Props#show'
-        # page_title gets picked up automatically by meta-tags gem
         @page_title = @property_details.title
         @page_description = @property_details.description
         # @page_keywords    = 'Site, Login, Members'
         return render "/pwb/props/show"
       else
+        @page_title = I18n.t("propertyNotFound")
+        hi_content = Content.where(tag: 'landing-carousel')[0]
+        @header_image = hi_content.present? ? hi_content.default_photo : nil
         return render "not_found"
       end
     end
@@ -31,10 +35,11 @@ module Pwb
       @operation_type = "for_sale"
       @operation_type_key = @operation_type.camelize(:lower)
       @property_details = Prop.find_by_id(params[:id])
+      @map_markers = []
 
       if @property_details && @property_details.visible && @property_details.for_sale
+        set_map_marker
         # gon.property_details =@property_details
-
         js property_details: @property_details
         js 'Pwb/Props#show'
         @page_title = @property_details.title
@@ -42,6 +47,9 @@ module Pwb
         # @page_keywords    = 'Site, Login, Members'
         return render "/pwb/props/show"
       else
+        @page_title = I18n.t("propertyNotFound")
+        hi_content = Content.where(tag: 'landing-carousel')[0]
+        @header_image = hi_content.present? ? hi_content.default_photo : nil
         return render "not_found"
       end
     end
@@ -69,7 +77,7 @@ module Pwb
                                origin_ip: request.ip,
                                user_agent: request.user_agent,
                                delivery_email: @current_agency.email_for_property_contact_form
-        # origin_email: params[:contact][:email]
+                               # origin_email: params[:contact][:email]
       })
 
       unless @enquiry.save && @client.save
@@ -97,5 +105,26 @@ module Pwb
       @error_messages = [I18n.t("contact.error"), e]
       return render "pwb/ajax/request_info_errors", layout: false
     end
+
+    private
+
+    def set_map_marker
+      if @property_details.show_map
+        @map_markers.push(
+          {
+            id: @property_details.id,
+            title: @property_details.title,
+            show_url: @property_details.contextual_show_path(@operation_type),
+            image_url: @property_details.primary_image_url,
+            display_price: @property_details.contextual_price_with_currency(@operation_type),
+            position: {
+              lat: @property_details.latitude,
+              lng: @property_details.longitude
+            }
+          }
+        )
+      end
+    end
+
   end
 end