+++ CHANGELOG.md
@@ -7,9 +7,13 @@ All notable changes to this project will be documented in this file.
 
 
 * Add Vue.js
+* Add search results map
+* Display infowindow on maps
 * Add social sharing buttons
-* Add Turkish and Vietnamese
+* Add Turkish and Vietnamese translations
 * Fix Rails 5.1.0 issue with seed task (#29)
+* Fix language picker redirecting to root
+* Set open graph meta tags
 * Support Rails 5.1.0
 
 ## 1.1.1 / 2017-07-25